20.10.19. index -> user 패키지 수정, menu_pic 입력 구현 중
