+++ Frogs_Matzip/src/main/java/com/frogs/matzip/Const.java
@@ -1,8 +1,10 @@
 package com.frogs.matzip;
 
 public class Const {
+	public static String realPath = null;
+	
 	public static final String TEMPLATE = "template";
-	public static final String HEADER = "tHeader";
+	public static final String HEADER = "header_temp";
 	public static final String VIEW = "view";
 	public static final String TITLE = "title";
 	public static final String CSS = "css";
