+++ Frogs_Matzip/src/main/java/com/frogs/matzip/IndexController.java
@@ -1,35 +1,30 @@
 package com.frogs.matzip;
 
-import javax.servlet.http.HttpSession;
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.frogs.matzip.model.IndexPARAM;
-import com.frogs.matzip.model.IndexVO;
+import com.frogs.matzip.rest.RestService;
 
 
 @Controller
 public class IndexController {
 	
 	@Autowired
-	private IndexService service;
-	
-	@RequestMapping(value="/logout", method = RequestMethod.GET) 
-	public String logout(HttpSession hs) {
-		hs.invalidate();
-		return "redirect:/";
-	}
+	private RestService restService;
 	
 	@RequestMapping(value="/index", method = RequestMethod.GET)
-	public String Index(Model model) {
+	public String Index(HttpServletRequest req, Model model) {
+		if(Const.realPath == null) {
+			Const.realPath = req.getSession().getServletContext().getRealPath("");
+		}
+		
 		// 인기 식당, 카테고리별 식당 리스트 뿌리기 해야 함
-		model.addAttribute("recRestList", service.selRecRestList());
+		model.addAttribute("recRestList", restService.selRecRestList());
 		
 		model.addAttribute(Const.CSS, new String[] {"common", "index", "indexheader"});
 		model.addAttribute(Const.TITLE, "FrogsMatzip");
@@ -37,28 +32,4 @@ public class IndexController {
 		model.addAttribute(Const.VIEW, "index");
 		return ViewRef.TEMP;
 	}
-	
-	@RequestMapping(value="/ajaxJoin", method = RequestMethod.POST)
-	@ResponseBody
-	public int ajaxJoin(@RequestBody IndexVO param) {
-		return service.ajaxJoin(param);
-	}
-	
-	@RequestMapping(value="/ajaxCheckId", method = RequestMethod.POST)
-	@ResponseBody
-	public int ajaxCheckId(@RequestBody IndexPARAM param) {
-		return service.login(param);
-	}
-	
-	@RequestMapping(value="ajaxLogin", method = RequestMethod.POST)
-	@ResponseBody
-	public int ajaxLogin(@RequestBody IndexPARAM param, HttpSession hs) {
-		int result = service.login(param);
-		if(result == Const.SUCCESS) {
-			hs.setAttribute(Const.LOGIN_USER, param);
-		}
-		return result;
-	}
-	
-	
 }
