--- Frogs_Matzip/src/main/java/com/frogs/matzip/model/IndexDMI.java
@@ -1,8 +1,0 @@
-package com.frogs.matzip.model;
-
-import org.apache.ibatis.type.Alias;
-
-@Alias("IndexDMI")
-public class IndexDMI extends IndexVO {
-	
-}
