--- Frogs_Matzip/src/main/java/com/frogs/matzip/model/IndexVO.java
@@ -1,72 +1,0 @@
-package com.frogs.matzip.model;
-
-import org.apache.ibatis.type.Alias;
-
-@Alias("IndexVO")
-public class IndexVO {
-	private int i_user;
-	private String user_id;
-	private String user_pw;
-	private String salt;
-	private String nm;
-	private String profile_img;
-	private String r_dt;
-	private String m_dt;
-	private int admin;
-	
-	
-	public int getAdmin() {
-		return admin;
-	}
-	public void setAdmin(int admin) {
-		this.admin = admin;
-	}
-	public int getI_user() {
-		return i_user;
-	}
-	public void setI_user(int i_user) {
-		this.i_user = i_user;
-	}
-	public String getUser_id() {
-		return user_id;
-	}
-	public void setUser_id(String user_id) {
-		this.user_id = user_id;
-	}
-	public String getUser_pw() {
-		return user_pw;
-	}
-	public void setUser_pw(String user_pw) {
-		this.user_pw = user_pw;
-	}
-	public String getSalt() {
-		return salt;
-	}
-	public void setSalt(String salt) {
-		this.salt = salt;
-	}
-	public String getNm() {
-		return nm;
-	}
-	public void setNm(String nm) {
-		this.nm = nm;
-	}
-	public String getProfile_img() {
-		return profile_img;
-	}
-	public void setProfile_img(String profile_img) {
-		this.profile_img = profile_img;
-	}
-	public String getR_dt() {
-		return r_dt;
-	}
-	public void setR_dt(String r_dt) {
-		this.r_dt = r_dt;
-	}
-	public String getM_dt() {
-		return m_dt;
-	}
-	public void setM_dt(String m_dt) {
-		this.m_dt = m_dt;
-	}
-}
