+++ Frogs_Matzip/src/main/java/com/frogs/matzip/user/model/UserDMI.java
@@ -1,0 +1,8 @@
+package com.frogs.matzip.user.model;
+
+import org.apache.ibatis.type.Alias;
+
+@Alias("UserDMI")
+public class UserDMI extends UserVO {
+
+}
